requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.2
Django==4.2.11
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
djangorestframework==3.15.1
pillow==10.2.0
pymemcache==4.0.0
sqlparse==0.4.4
typing_extensions==4.10.0
